+++ Source/Clients/Spreadsheets/GDataRowColumnCount.h
@@ -25,7 +25,7 @@
 // http://code.google.com/apis/spreadsheets/reference.html#gs_reference
 
 @interface GDataRowColumnCount : GDataObject <NSCopying> {
-  int count_;
+  NSInteger count_;
 }
 
 - (id)initWithXMLElement:(NSXMLElement *)element
@@ -33,20 +33,20 @@
 
 - (NSXMLElement *)XMLElement;
 
-- (int)count;
-- (void)setCount:(int)val;
+- (NSInteger)count;
+- (void)setCount:(NSInteger)val;
 
 @end
 
 @interface GDataColumnCount : GDataRowColumnCount <GDataExtension>
 
-+ (GDataColumnCount *)columnCountWithInt:(int)val;
++ (GDataColumnCount *)columnCountWithInt:(NSInteger)val;
 
 @end
 
 @interface GDataRowCount : GDataRowColumnCount <GDataExtension>
 
-+ (GDataRowCount *)rowCountWithInt:(int)val;
++ (GDataRowCount *)rowCountWithInt:(NSInteger)val;
 
 @end
